moved escapeHashtags to util
